+++ tests_host/test/protocol/test_protocol_cmd_led.c
@@ -41,6 +41,7 @@
 #include "protocol_cmd_adc.h"
 #include "protocol_cmd_eeprom.h"
 #include "protocol_cmd_led.h"
+#include "protocol_cmd_sensors.h"
 #include "protocol_cmd_sysinfo.h"
 #include "protocol_support.h"
 
@@ -51,6 +52,9 @@
 #include "mock_protocol_callback.h"
 #include "mock_adc.h"
 #include "mock_eeprom.h"
+#include "mock_lsm303accel.h"
+#include "mock_lsm303mag.h"
+#include "mock_sensors.h"
 
 uint32_t SystemCoreClock = 12000000; // overshadow the variable used to determine core lock
 static protMsgCommand_t message_cmd;